+++ source/renderer/app/api/utils/apiHelpers.ts
@@ -27,22 +27,21 @@ export const doesWalletRequireAdaToRemainToSupportTokens = (
   hasAssetsRemainingAfterTransaction?: boolean
 ): {
   requiresAdaToRemainToSupportNativeTokens: boolean;
-  adaToRemain?: number;
+  adaToProceed?: number;
 } => {
-  const adaToProceedRegex = new RegExp(
-    /.*I need approximately([\s\d.,]+)ada to proceed.*/
-  );
+  const adaToProceedRegex = /I need approximately\s+([\d.,]+)\s+ada to proceed/;
+
+  const [, adaToProceed] = adaToProceedRegex.exec(error.message) ?? [];
 
   if (
     error.code === 'cannot_cover_fee' &&
     hasAssetsRemainingAfterTransaction &&
-    adaToProceedRegex.test(error.message)
+    adaToProceed
   ) {
-    const roundedAda = Math.ceil(
-      Number(error.message.replace(adaToProceedRegex, '$1'))
-    );
-    const adaToRemain = roundedAda > 2 ? roundedAda : 2;
-    return { requiresAdaToRemainToSupportNativeTokens: true, adaToRemain };
+    return {
+      requiresAdaToRemainToSupportNativeTokens: true,
+      adaToProceed: Math.ceil(Number(adaToProceed)),
+    };
   }
   return { requiresAdaToRemainToSupportNativeTokens: false };
 };